backend/requirements.txt
fastapi 
uvicorn 
openai 
python-multipart 
pdfplumber 
python-dotenv